+++ pagetop/src/api/component/definition.rs
@@ -11,7 +11,7 @@ pub trait BaseComponent {
 
     fn single_name(&self) -> &'static str;
 
-    fn qualified_name(&self, last: u8) -> &'static str;
+    fn qualified_name(&self, last: usize) -> &'static str;
 }
 
 pub trait ComponentTrait: AnyComponent + BaseComponent + Send + Sync {
@@ -30,7 +30,7 @@ pub trait ComponentTrait: AnyComponent + BaseComponent + Send + Sync {
         true
     }
 
-    fn weight(&self) -> i8 {
+    fn weight(&self) -> isize {
         0
     }
 
@@ -57,7 +57,7 @@ impl<C: ?Sized + ComponentTrait> BaseComponent for C {
         util::partial_type_name(std::any::type_name::<Self>(), 1)
     }
 
-    fn qualified_name(&self, last: u8) -> &'static str {
+    fn qualified_name(&self, last: usize) -> &'static str {
         util::partial_type_name(std::any::type_name::<Self>(), last)
     }
 }
