+++ pagetop/src/base/component/container.rs
@@ -6,7 +6,7 @@ pub enum ContainerType { Header, Footer, Main, Section, Wrapper }
 
 pub struct Container {
     renderable   : fn() -> bool,
-    weight       : i8,
+    weight       : isize,
     components   : PageContainer,
     container    : ContainerType,
     id           : OptIden,
@@ -33,7 +33,7 @@ impl ComponentTrait for Container {
         (self.renderable)()
     }
 
-    fn weight(&self) -> i8 {
+    fn weight(&self) -> isize {
         self.weight
     }
 
@@ -127,7 +127,7 @@ impl Container {
         self
     }
 
-    pub fn with_weight(mut self, weight: i8) -> Self {
+    pub fn with_weight(mut self, weight: isize) -> Self {
         self.alter_weight(weight);
         self
     }
@@ -159,7 +159,7 @@ impl Container {
         self
     }
 
-    pub fn alter_weight(&mut self, weight: i8) -> &mut Self {
+    pub fn alter_weight(&mut self, weight: isize) -> &mut Self {
         self.weight = weight;
         self
     }
